requirements.txt
asgiref==3.5.2
Django==4.1
django-cleanup==6.0.0
django-cors-headers==3.13.0
django-quill-editor==0.1.40
gunicorn==20.1.0
sqlparse==0.4.2
tzdata==2022.2
whitenoise==6.2.0